test: refactor and reorganize functions
