+++ application_form/tests/conftest.py
@@ -1,10 +1,21 @@
 import faker.config
+import uuid
 from django.conf import settings
 from elasticsearch.helpers.test import get_test_client
 from elasticsearch_dsl.connections import add_connection
+from factory.faker import faker
 from pytest import fixture
 from rest_framework.test import APIClient
+from unittest.mock import Mock
 
+from apartment.tests.factories import IdentifierFactory
+from application_form.api.serializers import ApplicationSerializer
+from application_form.enums import ApplicationType
+from application_form.tests.factories import ApplicantFactory
+from application_form.tests.utils import (
+    calculate_ssn_suffix,
+    get_elastic_apartments_uuids,
+)
 from connections.tests.factories import ApartmentMinimalFactory
 
 faker.config.DEFAULT_LOCALE = "fi_FI"
@@ -62,3 +73,58 @@ def elastic_single_project_with_apartments(elasticsearch):
 
     for apartment in apartments:
         apartment.delete(refresh=True)
+
+
+def create_application_data(
+    profile, application_type=ApplicationType.HASO, num_applicants=2
+):
+    # Build apartments by creating identifiers
+    project_uuid, apartment_uuids = get_elastic_apartments_uuids()
+    apartments = IdentifierFactory.build_batch_for_att_schema(5, apartment_uuids)
+    apartments_data = [
+        {"priority": index, "identifier": apartment.identifier}
+        for index, apartment in enumerate(apartments)
+    ]
+    right_of_residence = 123456 if application_type == ApplicationType.HASO else None
+
+    # Build application request data
+    application_data = {
+        "application_uuid": str(uuid.uuid4()),
+        "application_type": application_type.value,
+        "ssn_suffix": profile.ssn_suffix,
+        "has_children": True,
+        "right_of_residence": right_of_residence,
+        "additional_applicant": None,
+        "project_id": str(project_uuid),
+        "apartments": apartments_data,
+    }
+    # Add a second applicant if needed
+    if num_applicants == 2:
+        date_of_birth = faker.Faker().date_of_birth(minimum_age=18)
+        applicant = ApplicantFactory.build()
+        application_data["additional_applicant"] = {
+            "first_name": applicant.first_name,
+            "last_name": applicant.last_name,
+            "email": applicant.email,
+            "street_address": applicant.street_address,
+            "postal_code": applicant.postal_code,
+            "city": applicant.city,
+            "phone_number": applicant.phone_number,
+            "date_of_birth": f"{date_of_birth:%Y-%m-%d}",
+            "ssn_suffix": calculate_ssn_suffix(date_of_birth),
+        }
+    return application_data
+
+
+def create_validated_application_data(
+    profile, application_type=ApplicationType.HASO, num_applicants=2
+):
+    application_data = create_application_data(
+        profile, application_type, num_applicants
+    )
+    serializer = ApplicationSerializer(
+        data=application_data,
+        context={"request": Mock(user=profile.user)},
+    )
+    serializer.is_valid(raise_exception=True)
+    return {**serializer.validated_data, "profile": profile}
