+++ application_form/tests/test_sales_application_api.py
@@ -3,7 +3,7 @@ from django.contrib.auth.models import Group
 from django.urls import reverse
 
 from application_form.models.application import Application
-from application_form.tests.utils import create_application_data
+from application_form.tests.conftest import create_application_data
 from users.enums import Roles
 from users.tests.factories import ProfileFactory
 from users.tests.utils import _create_token
